Publish lockfile as its own npm package.
