+++ __tests__/commands/install/lockfiles.js
@@ -4,7 +4,7 @@ import {run as check} from '../../../src/cli/commands/check.js';
 import * as constants from '../../../src/constants.js';
 import * as reporters from '../../../src/reporters/index.js';
 import {Install} from '../../../src/cli/commands/install.js';
-import Lockfile from '../../../src/lockfile/wrapper.js';
+import Lockfile from '../../../src/lockfile';
 import * as fs from '../../../src/util/fs.js';
 import {getPackageVersion, isPackagePresent, runInstall} from '../_helpers.js';
 import {promisify} from '../../../src/util/promise';
