+++ __tests__/package-hoister.js
@@ -2,7 +2,7 @@
 
 import PackageHoister, {HoistManifest} from '../src/package-hoister.js';
 import PackageResolver from '../src/package-resolver.js';
-import Lockfile from '../src/lockfile/wrapper.js';
+import Lockfile from '../src/lockfile';
 import type PackageReference from '../src/package-reference.js';
 import type Config from '../src/config.js';
 import type {Manifest} from '../src/types.js';
