+++ __tests__/package-resolver.js
@@ -3,7 +3,7 @@
 
 import * as reporters from '../src/reporters/index.js';
 import PackageResolver from '../src/package-resolver.js';
-import Lockfile from '../src/lockfile/wrapper.js';
+import Lockfile from '../src/lockfile';
 import Config from '../src/config.js';
 import makeTemp from './_temp.js';
 import * as fs from '../src/util/fs.js';
