+++ packages/lockfile/README.md
@@ -1,0 +1,18 @@
+# yarn-lockfile
+parse and/or write `yarn.lock` files
+
+## Usage Example
+
+```js
+const fs = require('fs')
+const lockfile = require('yarn-lockfile')
+
+let file = fs.readFileSync('yarn.lock', 'utf8')
+let json = lockfile.parse(file)
+
+console.log(json)
+
+let fileAgain = lockfile.stringify(json)
+
+console.log(fileAgain)
+```
