+++ src/cli/commands/install.js
@@ -6,12 +6,12 @@ import type {ReporterSelectOption} from '../../reporters/types.js';
 import type {Manifest, DependencyRequestPatterns} from '../../types.js';
 import type Config from '../../config.js';
 import type {RegistryNames} from '../../registries/index.js';
-import type {LockfileObject} from '../../lockfile/wrapper.js';
+import type {LockfileObject} from '../../lockfile';
 import normalizeManifest from '../../util/normalize-manifest/index.js';
 import {MessageError} from '../../errors.js';
 import InstallationIntegrityChecker from '../../integrity-checker.js';
-import Lockfile from '../../lockfile/wrapper.js';
-import lockStringify from '../../lockfile/stringify.js';
+import Lockfile from '../../lockfile';
+import {stringify as lockStringify} from '../../lockfile';
 import * as fetcher from '../../package-fetcher.js';
 import PackageInstallScripts from '../../package-install-scripts.js';
 import * as compatibility from '../../package-compatibility.js';
