+++ src/fetchers/git-fetcher.js
@@ -8,7 +8,7 @@ import * as fsUtil from '../util/fs.js';
 import * as constants from '../constants.js';
 import * as crypto from '../util/crypto.js';
 import {install} from '../cli/commands/install.js';
-import Lockfile from '../lockfile/wrapper.js';
+import Lockfile from '../lockfile';
 import Config from '../config.js';
 import {packTarball} from '../cli/commands/pack.js';
