+++ src/integrity-checker.js
@@ -1,7 +1,8 @@
 /* @flow */
 
 import type Config from './config.js';
-import type {LockManifest} from './lockfile/wrapper.js';
+import type {LockManifest} from './lockfile';
+import type {RegistryNames} from './registries/index.js';
 import * as constants from './constants.js';
 import * as fs from './util/fs.js';
 import {sortAlpha, compareSortedArrays} from './util/misc.js';
