+++ src/lockfile/parse.js
@@ -42,7 +42,7 @@ function isValidPropValueToken(token): boolean {
   return VALID_PROP_VALUE_TOKENS.indexOf(token.type) >= 0;
 }
 
-export function* tokenise(input: string): Iterator<Token> {
+function* tokenise(input: string): Iterator<Token> {
   let lastNewline = false;
   let line = 1;
   let col = 0;
@@ -160,7 +160,7 @@ export function* tokenise(input: string): Iterator<Token> {
   yield buildToken(TOKEN_TYPES.eof);
 }
 
-export class Parser {
+class Parser {
   constructor(input: string, fileLoc: string = 'lockfile') {
     this.comments = [];
     this.tokens = tokenise(input);
